requirements.txt
# ── Core Framework ──────────────────────────────────────────
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

# ── Flask Extensions ────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
flask-cors==6.0.2

# ── Database / ORM ──────────────────────────────────────────
SQLAlchemy==2.0.47
alembic==1.18.4
Mako==1.3.10
greenlet==3.3.2

# ── Auth & Security ─────────────────────────────────────────
bcrypt==5.0.0
PyJWT==2.11.0

# ── Environment / Config ───────────────────────────────────
python-dotenv==1.2.1

# ── Database Drivers ────────────────────────────────────────
psycopg2-binary==2.9.11          # PostgreSQL (production / Docker)
# SQLite is used by default for local dev (no extra driver needed)

# ── AI / NLP ──────────────────────────────────────────────────
scikit-learn>=1.4                # TF-IDF + cosine similarity for AI assistant intent matching

# ── Platform / Typing ──────────────────────────────────────
typing_extensions==4.15.0
colorama==0.4.6                  # colored CLI output (Windows)

# ── Testing ─────────────────────────────────────────────────
pytest==8.3.0
pytest-flask==1.3.0
